Separate base models from generic
