+++ routers/getconfig.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter
 from fastapi.encoders import jsonable_encoder
 
 # load models
-from backend.core.models.models import GetConfig
+from backend.core.models.generic_models import GetConfig
 from backend.core.models.napalm import NapalmGetConfig
 from backend.core.models.ncclient import NcclientGetConfig
 from backend.core.models.netmiko import NetmikoGetConfig
